# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AnimusForge API
  description: |
    ## Adaptive AI Persona Framework with Ethical Governance

    AnimusForge is an enterprise-grade framework for creating, managing, and evolving
    AI personas with built-in ethical governance, kill-switch mechanisms, and
    comprehensive memory systems.

    ### Key Features

    - **Persona Management**: Create, update, activate, and evolve AI personas with
      configurable traits, cognitive patterns, and behavioral constraints.
    - **LLM Gateway**: Multi-provider LLM access with automatic failover, load balancing,
      and cost optimization across OpenAI, Anthropic, DeepSeek, and GLM.
    - **Kill-Switch**: Emergency stop and recovery mechanisms for immediate persona
      deactivation with audit trail and recovery workflows.
    - **Memory System**: Vector-based semantic memory and graph-based knowledge storage
      for persistent persona context and learning.
    - **Ethics Engine**: Real-time content evaluation, audit logging, and compliance
      reporting across multiple ethical categories.
    - **System Monitoring**: Health checks, Prometheus metrics, and configuration
      management endpoints.

    ### Authentication

    JWT authentication is supported. Include `Authorization: Bearer <token>` header
    for authenticated requests. Some endpoints may require elevated permissions.

    ### Rate Limiting

    All endpoints are rate-limited. Check response headers:
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-03-06T12:00:00Z"
...
    EthicsCategory:
      type: string
      enum:
        - harm
        - fairness
        - privacy
        - transparency
        - accountability
        - benevolence
        - autonomy
        - truthfulness
    EthicsEvaluateRequest:
      type: object
      required:
        - persona_id
        - content
        - content_type
      properties:
        persona_id:
          $ref: "#/components/schemas/UUID"
        content:
          type: string
          minLength: 1
        content_type:
          type: string
          enum: [text, code, interaction, decision, output, behavior]
        categories:
          type: array
          items:
            $ref: "#/components/schemas/EthicsCategory"
        strict_mode:
          type: boolean
          default: false
    EthicsEvaluateResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        persona_id:
          $ref: "#/components/schemas/UUID"
        compliant:
          type: boolean
        overall_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        action:
          type: string
          enum: [allow, warn, modify, block, escalate]
        violations:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        evaluated_at:
          $ref: "#/components/schemas/Timestamp"

    # ==================== System Schemas ====================
...
paths:
...
  /api/v1/ethics/evaluate:
    post:
      tags:
        - Ethics
      summary: Evaluate content
      description: Evaluate content for ethics compliance across multiple categories.
      operationId: evaluateContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EthicsEvaluateRequest"
      responses:
        "200":
          description: Evaluation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EthicsEvaluateResponse"
...
  /api/v1/ethics/audit:
    get:
      tags:
        - Ethics
      summary: List audits
      description: Retrieve a paginated list of audit entries.
      operationId: listAudits
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: persona_id
          in: query
          schema:
            $ref: "#/components/schemas/UUID"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/EthicsCategory"
      responses:
        "200":
          description: List of audit entries
  /api/v1/ethics/violations:
    get:
      tags:
        - Ethics
      summary: List violations
      description: Retrieve a paginated list of ethics violations.
      operationId: listViolations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: persona_id
          in: query
          schema:
            $ref: "#/components/schemas/UUID"
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, low, medium, high, critical]
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of violations
  /api/v1/ethics/report:
    post:
      tags:
        - Ethics
      summary: Generate report
      description: Generate a comprehensive ethics report.
      operationId: generateEthicsReport
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                persona_id:
                  $ref: "#/components/schemas/UUID"
                start_date:
                  $ref: "#/components/schemas/Timestamp"
                end_date:
                  $ref: "#/components/schemas/Timestamp"
                include_violations:
                  type: boolean
                  default: true
                include_audits:
                  type: boolean
                  default: true
                format:
                  type: string
                  enum: [json, pdf, html]
                  default: json
      responses:
        "200":
          description: Ethics report generated

  # ==================== System ====================
...